arthur/app/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
black==25.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
libvirt-python==11.6.0
MarkupSafe==2.1.5
minio==7.2.8
pycparser==2.22
pycryptodome==3.20.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.5
rpds-py==0.20.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
pika==1.3.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.39
pydantic==2.11.7
pydantic-settings==2.10.1
GitPython==3.1.44
pytest
requests==2.32.5
pytest-cov